+++ src/main/java/com/foolsix/fancyenchantments/enchantment/util/EnchUtils.java
@@ -9,6 +9,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Style;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -238,5 +239,25 @@ public class EnchUtils {
         if (instance != null) {
             living.forceAddEffect(new MobEffectInstance(effect, instance.getDuration(), amplifier), null);
         }
+    }
+
+    public static boolean canBlock(LivingEntity holder, @Nullable Vec3 sourcePosition) {
+        // source position should never be null (checked by livingentity) but safety as its marked nullable
+        if (sourcePosition == null) {
+            return false;
+        }
+        float blockAngle = 45;
+        // want the angle between the view vector and the
+        Vec3 viewVector = holder.getViewVector(1.0f);
+        Vec3 entityPosition = holder.position();
+        Vec3 direction = new Vec3(entityPosition.x - sourcePosition.x, 0, entityPosition.z - sourcePosition.z);
+        double length = viewVector.length() * direction.length();
+        // prevent zero vector from messing with us
+        if (length < 1.0E-4D) {
+            return false;
+        }
+        // acos will return between 90 and 270, we want an absolute angle from 0 to 180
+        double angle = Math.abs(180 - Math.acos(direction.dot(viewVector) / length) * Mth.RAD_TO_DEG);
+        return blockAngle >= angle;
     }
 }